fix: int 최댓값 입력시 overflow로 인한 음수 결과 안나오도록 수정

+++ src/main/java/calculator/Calculator.java
@@ -22,7 +22,7 @@ public class Calculator {
     public long addInputNums(String input) {
         List<Integer> inputNums = inputExtractor.getNumbers(input);
         return inputNums.stream()
-                .mapToInt(Integer::intValue)
+                .mapToLong(Integer::intValue)
                 .sum();
     }
 }
